requirements.txt
alpaca-py
pandas
numpy
matplotlib
streamlit
python-dotenv
plotly
yfinance
